+++ equal.ts
@@ -1,5 +1,5 @@
 // Copyright 2021-present the Equal authors. All rights reserved. MIT license.
-import { and, AnyFn, entries, F, has, length, N, xor } from "./deps.ts";
+import { and, AnyFn, entries, F, has, ifElse, length, N, xor } from "./deps.ts";
 import {
   isBothArray,
   isBothDate,
@@ -11,6 +11,7 @@ import {
   isBothObjectExcludeJSON,
   isBothPrimitive,
   isBothRegExp,
+  isBothSet,
 } from "./is.ts";
 import { entriesSymbol, instanceOf } from "./utils.ts";
 import { is } from "./constants.ts";
@@ -49,6 +50,7 @@ const equal = <T, U extends T>(a: T, b: U): boolean => {
     [isBothRegExp, equalRegExp],
     [isBothError, equalError],
     [isBothMap, equalMap],
+    [isBothSet, equalSet],
     [isBothObjectExcludeJSON, equalObjectExcludeJson],
   ];
 
@@ -60,14 +62,40 @@ const equal = <T, U extends T>(a: T, b: U): boolean => {
   return false;
 };
 
+const equalConstructor = <T, U extends T>(
+  obj: Function,
+  a: T,
+  b: U,
+): boolean => and(instanceOf(obj, a), () => instanceOf(obj, b));
+
 const equalRegExp = <T extends RegExp, U extends T>(a: T, b: U): boolean =>
   a.toString() === b.toString();
 
 const equalDate = <T extends Date, U extends T>(a: T, b: U): boolean =>
   a.getTime() === b.getTime();
 
-const equalError = <T extends Error, U extends T>(a: T, b: U): boolean =>
-  and(a.message === b.message, () => a.toString() === b.toString());
+const equalError = <T extends Error, U extends T>(a: T, b: U): boolean => {
+  if (a.message !== b.message) return false;
+  const errorConstructors = [
+    EvalError,
+    RangeError,
+    ReferenceError,
+    SyntaxError,
+    TypeError,
+    URIError,
+  ];
+  if (
+    errorConstructors.some((constructor) => equalConstructor(constructor, a, b))
+  ) {
+    return true;
+  }
+
+  return ifElse(equalConstructor(AggregateError, a, b), () =>
+    equalArray(
+      (a as Error as AggregateError).errors,
+      (b as Error as AggregateError).errors,
+    ), () => a.constructor.name === b.constructor.name);
+};
 
 const equalFunction = <T extends Function, U extends T>(a: T, b: U): boolean =>
   a.toString() === b.toString();
@@ -80,6 +108,11 @@ const equalMap = <T extends Map<any, any>, U extends T>(
 
   return equalKeyValueTupleNoOrder([...a], [...b]);
 };
+
+const equalSet = <T extends Set<unknown>, U extends T>(
+  a: T,
+  b: U,
+): boolean => equalArray([...a], [...b]);
 
 const equalKeyValueTuple = <T extends [unknown, unknown], U extends T>(
   [keyA, valueA]: T,
@@ -119,9 +152,7 @@ const equalObjectExcludeJson = <
   b: U,
 ): boolean => {
   if (
-    [Number, String, Boolean].some((obj) =>
-      and(instanceOf(obj, a), () => instanceOf(obj, b))
-    )
+    [Number, String, Boolean].some((obj) => equalConstructor(obj, a, b))
   ) {
     return equal(a.valueOf(), b.valueOf());
   }
@@ -141,6 +172,7 @@ const equalArray = <T extends unknown[], U extends T>(a: T, b: U): boolean => {
 export {
   equal,
   equalArray,
+  equalConstructor,
   equalDate,
   equalError,
   equalFunction,
@@ -150,4 +182,5 @@ export {
   equalMap,
   equalObjectExcludeJson,
   equalRegExp,
+  equalSet,
 };