+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# [v1.2.0-beta.3](https://github.com/TomokiMiyauci/equal/compare/v1.2.0-beta.2...v1.2.0-beta.3) (2021-05-25)
+
+## ✨ New Features
+- [`db6f997`](https://github.com/TomokiMiyauci/equal/commit/db6f997)   Add equality of AggregateError (Issues: [`#9`](https://github.com/TomokiMiyauci/equal/issues/9))
+
+# [v1.2.0-beta.2](https://github.com/TomokiMiyauci/equal/compare/v1.2.0-beta.1...v1.2.0-beta.2) (2021-05-25)
+
+## ✨ New Features
+- [`6f6a490`](https://github.com/TomokiMiyauci/equal/commit/6f6a490)   Add feature of equality all built-in Errors (Issues: [`#9`](https://github.com/TomokiMiyauci/equal/issues/9))
+
+# [v1.2.0-beta.1](https://github.com/TomokiMiyauci/equal/compare/v1.1.0...v1.2.0-beta.1) (2021-05-25)
+
+## ✨ New Features
+- [`60874d1`](https://github.com/TomokiMiyauci/equal/commit/60874d1)   Add a feature of judge Set object correctly (Issues: [`#8`](https://github.com/TomokiMiyauci/equal/issues/8))
+
 # [v1.1.0](https://github.com/TomokiMiyauci/equal/compare/v1.0.2...v1.1.0) (2021-05-24)
 
 ## ✨ New Features